requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
Django==4.2
django-redis==5.4.0
django-rest-framework-docs==0.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
jsonpickle==3.0.2
packaging==23.2
pillow==10.2.0
psycopg2==2.9.9
PyJWT==1.7.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
